CORE-8400 Refactor Preferences to have separate view and presenter
