+++ de-lib/src/main/java/org/iplantc/de/desktop/Desktop.gwt.xml
@@ -33,6 +33,7 @@
     <inherits name="org.iplantc.de.notifications.Notifications"/>
     <inherits name="org.iplantc.de.systemMessages.SystemMessages"/>
     <inherits name="org.iplantc.de.collaborators.Collaborators" />
+    <inherits name="org.iplantc.de.preferences.Preferences"/>
 
     <source path="client"/>
     <source path="shared"/>
