+++ de-lib/src/main/java/org/iplantc/de/desktop/client/DesktopView.java
@@ -209,6 +209,10 @@ public interface DesktopView extends IsWidget {
 
         void stickWindowToTop(Window window);
 
+        void doPeriodicSessionSave();
+
+        void restoreWindows(List<WindowState> windowStates);
+
         void setPeriodicSessionFailFlags();
     }
