+++ de-lib/src/main/java/org/iplantc/de/preferences/Preferences.gwt.xml
@@ -1,0 +1,13 @@
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.0//EN"
+        "http://google-web-toolkit.googlecode.com/svn/releases/2.0/distro-source/core/src/gwt-module.dtd">
+<module>
+
+    <inherits name='com.google.gwt.user.User'/>
+    <inherits name="org.iplantc.de.DiscoveryEnvironmentCommon"/>
+    <inherits name="org.iplantc.de.commons.ui-commons" />
+    <inherits name="org.iplantc.de.apps.widgets.Widgets"/>
+
+    <source path="client"/>
+    <source path="shared"/>
+
+</module>
