+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/desktop/Desktop.gwt.xml
@@ -38,7 +38,7 @@
     </replace-with>
 
     <replace-with class="org.iplantc.de.theme.base.client.desktop.PreferencesViewDefaultAppearance">
-        <when-type-is class="org.iplantc.de.desktop.client.views.widgets.PreferencesDialog.PreferencesViewAppearance" />
+        <when-type-is class="org.iplantc.de.preferences.client.PreferencesView.PreferencesViewAppearance"/>
     </replace-with>
 
     <replace-with class="org.iplantc.de.theme.base.client.desktop.presenter.DesktopPresenterDefaultAppearance">
