+++ de-webapp/src/main/java/org/iplantc/de/client/gin/DEInjector.java
@@ -10,6 +10,7 @@ import org.iplantc.de.desktop.client.gin.DEGinModule;
 import org.iplantc.de.diskResource.client.gin.DiskResourceGinModule;
 import org.iplantc.de.fileViewers.client.gin.FileViewerGinModule;
 import org.iplantc.de.notifications.client.gin.NotificationGinModule;
+import org.iplantc.de.preferences.client.gin.PreferencesGinModule;
 import org.iplantc.de.tags.client.gin.TagsGinModule;
 import org.iplantc.de.tools.requests.client.gin.ToolRequestGinModule;
 
@@ -31,7 +32,8 @@ import com.google.gwt.inject.client.Ginjector;
                 CommentsGinModule.class,
                 TagsGinModule.class,
                 FileViewerGinModule.class,
-                NotificationGinModule.class})
+                NotificationGinModule.class,
+                PreferencesGinModule.class})
 public interface DEInjector extends Ginjector {
     public static final DEInjector INSTANCE = GWT.create(DEInjector.class);
